requirements.txt
Flask>=2.3.0
gunicorn>=20.1.0
flask-cors>=4.0.0
Werkzeug>=2.3.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-docx>=0.8.11
pypdf2>=3.0.0
Flask-Login>=0.6.2
Flask-Bcrypt>=1.0.1